contrib/apmsql: fix build for Go 1.8/1.9
Move some bits out to files with version build tags
to support tracing newer versions of database/sql
without breaking support for older versions of Go.
