+++ contrib/apmsql/conn_pre_go110.go
@@ -1,0 +1,9 @@
+// +build !go1.10
+
+package apmsql
+
+import "database/sql/driver"
+
+type connGo110 struct{}
+
+func (connGo110) init(in driver.Conn) {}
