+++ contrib/apmsql/driver.go
@@ -1,7 +1,6 @@
 package apmsql
 
 import (
-	"context"
 	"database/sql"
 	"database/sql/driver"
 	"errors"
@@ -10,7 +9,6 @@ import (
 	"strings"
 
 	"github.com/elastic/apm-agent-go/contrib/apmsql/dsn"
-	"github.com/elastic/apm-agent-go/trace"
 )
 
 // DriverPrefix should be used as a driver name prefix when
@@ -100,44 +98,8 @@ func (d *tracingDriver) querySignature(query string) string {
 	return strings.ToUpper(fields[0])
 }
 
-func (d *tracingDriver) OpenConnector(name string) (driver.Connector, error) {
-	if dc, ok := d.Driver.(driver.DriverContext); ok {
-		oc, err := dc.OpenConnector(name)
-		if err != nil {
-			return nil, err
-		}
-		return &driverConnector{oc.Connect, d, name}, nil
-	}
-	connect := func(context.Context) (driver.Conn, error) {
-		return d.Driver.Open(name)
-	}
-	return &driverConnector{connect, d, name}, nil
-}
-
 func (d *tracingDriver) Open(name string) (driver.Conn, error) {
 	return nil, errors.New("Open should not be called")
-}
-
-type driverConnector struct {
-	connect func(context.Context) (driver.Conn, error)
-	driver  *tracingDriver
-	name    string
-}
-
-func (d *driverConnector) Connect(ctx context.Context) (driver.Conn, error) {
-	span, ctx := trace.StartSpan(ctx, "connect", d.driver.spanType("connect"))
-	if span != nil {
-		defer span.Done(-1)
-	}
-	conn, err := d.connect(ctx)
-	if err != nil {
-		return nil, err
-	}
-	return newConn(conn, d.driver, d.name), nil
-}
-
-func (d *driverConnector) Driver() driver.Driver {
-	return d.driver
 }
 
 func driverName(d driver.Driver) string {
