+++ contrib/apmsql/stmt_pre_go19.go
@@ -1,0 +1,9 @@
+// +build !go1.9
+
+package apmsql
+
+import "database/sql/driver"
+
+type stmtGo19 struct{}
+
+func (stmtGo19) init(in driver.Stmt) {}
